Fix java-http-client tests
